requirements.txt
ccxt>=4.2
pandas>=2.0
scikit-learn>=1.4
joblib>=1.3
python-binance>=1.0.19
matplotlib>=3.8
numpy>=1.26
